[back][#10] clean code, deplace initWebSocket in file friends.js for receive invitation anywhere
